% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.job.ad.chart-package.R
\docType{package}
\name{ons.job.ad.chart-package}
\alias{ons.job.ad.chart}
\alias{ons.job.ad.chart-package}
\title{ons.job.ad.chart: Create chart of job adverts}
\description{
Code downloads job advert data via ONS API then saves a chart of indexed values for 2022 for all industries and healthcare.
}
\author{
\strong{Maintainer}: Matthew Katz \email{matthew.katz@dhsc.gov.uk}

}
\keyword{internal}
