% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Create chart of job adverts}
\usage{
run_analysis()
}
\description{
Code downloads job advert data via ONS API then
saves a chart of indexed values for 2022 for all
industries and healthcare.
}
